refactors code

+++ src/apis/comments.ts
@@ -9,7 +9,7 @@ export const CreatePostMutation = gql`
 `
 
 export const GetCommentsQuery = gql`
-  query GetComment($postId: Float!, $page: Int!, $itemsPerPage: Int!) {
+  query GetComment($postId: Int!, $page: Int!, $itemsPerPage: Int!) {
     getComment(postId: $postId, page: $page, itemsPerPage: $itemsPerPage) {
       id
       text
@@ -34,7 +34,7 @@ export const GetCommentsQuery = gql`
 `
 
 export const GetCommentQuery = gql`
-  query Comment($commentId: Float!) {
+  query Comment($commentId: Int!) {
     Comment(postId: $commentId) {
       id
       text
@@ -58,7 +58,7 @@ export const GetCommentQuery = gql`
   }
 `
 export const GetCommentsRepliesQuery = gql`
-  query GetRepliesOfComment($commentId: Float!, $page: Int!, $itemsPerPage: Int!) {
+  query GetRepliesOfComment($commentId: Int!, $page: Int!, $itemsPerPage: Int!) {
     getRepliesOfComment(commentId: $commentId, page: $page, itemsPerPage: $itemsPerPage) {
       id
       text
@@ -83,7 +83,7 @@ export const GetCommentsRepliesQuery = gql`
 `
 
 export const AddCommentMutation = gql`
-  mutation AddComment($text: String!, $postId: Float!) {
+  mutation AddComment($text: String!, $postId: Int!) {
     createComment(data: { text: $text, postId: $postId }) {
       id
       text
@@ -108,7 +108,7 @@ export const AddCommentMutation = gql`
 `
 
 export const AddReplyToCommentMutation = gql`
-  mutation AddReplyToComment($text: String!, $postId: Float!, $commentId: Float!) {
+  mutation AddReplyToComment($text: String!, $postId: Int!, $commentId: Int!) {
     addReplyToComment(data: { text: $text, postId: $postId, commentId: $commentId }) {
       id
       text
